Changed vars to int64

+++ frametable.h
@@ -2,6 +2,6 @@
 #define FRAMETABLE
 #include "hardware.h"
 
-int checkFrameTable (long pa, struct Process *process, struct Hardware *hardware);
-int allocateNewFrame(struct Process *process, struct Hardware *hardware);
+int64_t checkFrameTable (int64_t pa, struct Process *process, struct Hardware *hardware);
+int64_t allocateNewFrame(struct Process *process, struct Hardware *hardware);
 #endif
