+++ hardware.h
@@ -4,23 +4,23 @@
 #include <inttypes.h>
 
 struct Node {
-	short index;
+	int64_t index;
 	struct Node *next;
 };
 
 struct L2CacheSet
 {
-	unsigned int tags[8];
-	unsigned int dirty[8];
-	unsigned int valid[8];
+	int64_t tags[8];
+	int64_t dirty[8];
+	int64_t valid[8];
 	struct Node *head;
 	struct Node *tail;
-	short listCount;
+	int64_t listCount;
 };
 
 struct L1Cache
 {
-	unsigned int tags[128];
+	int64_t tags[128];
 };
 
 struct L2Cache
@@ -30,39 +30,39 @@ struct L2Cache
 
 struct VictimCache
 {
-	unsigned int tags[8];
-	unsigned short lruCounter[8];
-	unsigned short valid[8];
+	int64_t tags[8];
+	int64_t lruCounter[8];
+	int64_t valid[8];
 };
 
 struct FrameTable
 {
-	int pid[65536];
-	int initialFrameAlloc;
+	int64_t pid[65536];
+	int64_t initialFrameAlloc;
 };
 
 struct PageTable
 {
-	unsigned short entries[256];
+	int64_t entries[256];
 };
 
 struct MainMemory
 {
-	unsigned int freeFrames;
-	unsigned short nextFreeFrame;
+	int64_t freeFrames;
+	int64_t nextFreeFrame;
 	struct PageTable* frames[65536];
 
-	uint16_t lru[65536];
+	int64_t lru[65536];
 
 };
 
 struct TLB
 {
-	long pageNumber[32];
-	long frameNumber[32];
-	int pid[32];
-	unsigned int valid[32];
-	short lru[32][32];
+	int64_t pageNumber[32];
+	int64_t frameNumber[32];
+	int64_t pid[32];
+	int64_t valid[32];
+	int64_t lru[32][32];
 };
 
 struct Hardware
@@ -77,18 +77,18 @@ struct Hardware
 
 struct SegmentTable
 {
-	unsigned short csBase;
-	unsigned short dsBase;
-	unsigned int csLength;
-	unsigned int dsLength;
+	int64_t csBase;
+	int64_t dsBase;
+	int64_t csLength;
+	int64_t dsLength;
 };
 
 struct Process
 {
 
-	int pid;
+	int64_t pid;
 	struct SegmentTable *ldt;
-	short state; // Thrashing: 0 -> suspended, 1 -> active
+	int64_t state; // Thrashing: 0 -> suspended, 1 -> active
 };
 
 #endif
