+++ init.h
@@ -4,4 +4,4 @@ void victimCacheInit(struct Hardware *hardware);
 void l2CacheInit(struct Hardware *hardware);
 void frameTableInit(struct Hardware *hardware);
 void mainMemoryInit(struct Hardware *hardware);
-struct Process *processInit(int pid);
+struct Process *processInit(int64_t pid);
