+++ tlb.c
@@ -3,27 +3,27 @@
 #include "hardware.h"
 #include "tlb.h"
 
-long fetchTLB(struct Hardware *hardware, long pageNumber, int pid) {
+int64_t fetchTLB(struct Hardware *hardware, int64_t pageNumber, int64_t pid) {
     printf("Checking in TLB...\n");
     for (int i=0;i<32;i++) {
         if(hardware->tlb->valid[i] && hardware->tlb->pid[i] == pid && hardware->tlb->pageNumber[i] == pageNumber) {
-            printf("TLB hit for Process: %d & Page Number: %ld\n", pid,pageNumber);            
+            printf("TLB hit for Process: %d & Page Number: %lld\n", pid,pageNumber);            
             updateMatrix(hardware, i); // updating LRU square matrix
             return hardware->tlb->frameNumber[i];
         }
     }
-    printf("TLB miss for Process: %d & Page Number: %ld\n", pid,pageNumber);
+    printf("TLB miss for Process: %d & Page Number: %lld\n", pid,pageNumber);
     return -1; // tlb miss
 }
 
-void invalidateTLB(struct Hardware *hardware, int pid) {
+void invalidateTLB(struct Hardware *hardware, int64_t pid) {
     for(int i=0;i<32;i++) {
         if(hardware->tlb->pid[i] == pid)
             hardware->tlb->valid[i] = 0;
     }
 }
 
-void updateTLB(struct Hardware *hardware, long pageNumber, long frameNumber, int pid) {
+void updateTLB(struct Hardware *hardware, int64_t pageNumber, int64_t frameNumber, int64_t pid) {
     for(int i=0;i<32;i++) {
         if(!hardware->tlb->valid[i]) {
             printf("Found an invalid line!\nUpdating TLB...\n");
@@ -46,7 +46,7 @@ void updateTLB(struct Hardware *hardware, long pageNumber, long frameNumber, int
                 zeroCount++;
         }
         if(zeroCount == 32) {
-            printf("Found a line to evict, Process: %d, Page Number: %ld & Frame Number: %ld\n",hardware->tlb->pid[i], hardware->tlb->pageNumber[i], hardware->tlb->frameNumber[i]);
+            printf("Found a line to evict, Process: %d, Page Number: %lld & Frame Number: %lld\n",hardware->tlb->pid[i], hardware->tlb->pageNumber[i], hardware->tlb->frameNumber[i]);
             hardware->tlb->pid[i] = pid;
             hardware->tlb->pageNumber[i] = pageNumber;
             hardware->tlb->frameNumber[i] = frameNumber;
@@ -59,7 +59,7 @@ void updateTLB(struct Hardware *hardware, long pageNumber, long frameNumber, int
     return;
 }
 
-void updateMatrix(struct Hardware *hardware, int line) { 
+void updateMatrix(struct Hardware *hardware, int64_t line) { 
     for(int j=0;j<32;j++)
         hardware->tlb->lru[line][j] = 1;
     for(int j=0;j<32;j++)
@@ -67,12 +67,12 @@ void updateMatrix(struct Hardware *hardware, int line) {
     return;
 }
 
-void invalidateLine(struct Hardware *hardware, long pageNumber, int pid) {
+void invalidateLine(struct Hardware *hardware, int64_t pageNumber, int64_t pid) {
     for(int i=0;i<32;i++) {
         if(hardware->tlb->pid[i] == pid && hardware->tlb->pageNumber[i] == pageNumber) {
             printf("Found the line in TLB to invalidate!\n");
             hardware->tlb->valid[i] = 0;
-            printf("Invalidated the line with Process: %d and Page Number: %ld\n", pid, pageNumber);
+            printf("Invalidated the line with Process: %d and Page Number: %lld\n", pid, pageNumber);
             return;
         }
     }
